refactor(search): ChunkSearchResult の score を int から float に広げる (#384) (#385)
nene-recall ADR 0006（2026-09-01 起草）の先行依頼。指示書 =
_work/handoff-corpus-2026-09-01-score-float-work-order.md・board L132（due 2026-09-12）。

■ 値は変わらない
PdoChunkSearchRepository が返すのは CASE WHEN ... THEN 1 ELSE 0 END の合計＝
整数値そのもの。float に広げても 1 が 1.0 になるだけで、振る舞いの変化はゼロ。
「無風で通せる窓が今」というのが本便の前提。

■ なぜ差し替えの前にやるか
nene-recall（Go 製ハイブリッド検索）を検索バックエンドへ差し替える構想が Phase 2 に
ある。ベクトル類似度は float なので、差し替え時点で int が必ず障壁になる。
差し替えと同時にやると型変更とバックエンド交換が同じ PR に乗り、検索結果が変わった
とき「型のせいか、バックエンドのせいか」の切り分けができなくなる。

■ 変更（6箇所・実測した全数）
- src/Search/ChunkSearchResult.php:13         public int $score -> public float $score
- src/Search/PdoChunkSearchRepository.php:83  (int) -> (float)
- tests/Search/PdoChunkSearchRepositoryTest.php:92/114/116  assertSame(1/2/1) -> 1.0/2.0/1.0
- tests/Search/SearchChunksUseCaseTest.php:38 score: 1 -> score: 1.0

残る5行（PdoChunkSearchRepository.php:46,54,66,70,75）は $scoreParts / relevance_score
の列名と SQL 組み立てで、型に触らない。assertCount の 1 / 2 は件数であって score では
ないので触っていない。

■ assertSame のまま直した（assertEquals へ逃げていない）
PHPUnit の assertSame は === 比較なので 1.0 === 1 は false。float 化で3件落ちるのは
想定内であり、型が実際に広がった証拠でもある。緩い比較へ逃げると次に float が壊れても
気づけない。裏を返すと、更新後のテストが 1.0 で通ること自体が型が広がった証拠になる。

■ 外部契約は変わらない（実測）
score が HTTP 応答の JSON に出る経路は 0 件。frontend/src と docs/openapi への grep が
いずれも 0 ヒット。OpenAPI の変更は不要で、API の破壊的変更ではない。

■ 触っていないもの
検索ロジック（LIKE の部分一致スコア）には一切触っていない。それを置き換えるのが Recall
の仕事で、本便のスコープは型だけ。ここを一緒に直すと「値が変わらない」という本便の
最大の利点が消える。SearchChunksUseCase / ChunkSearchRepositoryInterface のシグネチャも
変えていない（Phase 2 の差し替えはこの2つを無傷で行える設計）。

品質ゲート = composer check 全通過（PHPUnit 425/425・PHPStan level8 no errors・
CS-Fixer・OpenAPI・MCP・conformance）。

+++ src/Search/ChunkSearchResult.php
@@ -10,7 +10,7 @@ final readonly class ChunkSearchResult
 {
     public function __construct(
         public Chunk $chunk,
-        public int $score,
+        public float $score,
     ) {
     }
 }
